feat: complete UI for watched until
